requirements.txt
-r requirements_core.txt

# The following packages are used exclusively in the production configuration.
# 
# Use requirements_core.txt for local development:
# 
#     pip install -r requirements_core.txt

gevent==1.0.1
greenlet==0.4.2
boto==2.32.1
dj-database-url==0.3.0
gunicorn==19.1.0
psycopg2==2.5.3
django-storages==1.1.8
